+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 #### 22 February 2024
 
 * Load client configuration from file instead of hardcoding the values.
+* Move client connection settings to the configuration file.
 
 #### 21 February 2024